frontend

+++ src/components/header/useCloseSession.jsx
@@ -2,22 +2,34 @@ import { useNavigate} from 'react-router-dom'
 import {useState,useContext} from 'react'
 import fs from '../../services/TokenService'
 import Context from '../../provider/Context'
+import useHttpPost from '../../hooks/useHttpPost' 
+import {useEffect} from 'react' 
 
 const useCloseSession = () => {
   const navigate = useNavigate()
   const {setUser} = useContext(Context)
+  const {postData,errors,data,loading,loadingAdd} = useHttpPost()
+
+  useEffect(() => {
+    if (loadingAdd===true) {
+
+      fs.deleteToken();
+
+      setTimeout(() => {
+
+        setUser(null)
+        navigate('/pt/login')
+
+      },1000)
+
+    }
+  },[loadingAdd])
   
   const handleOnClickCloseSession = () => {
-    fs.deleteToken();
-
-    setTimeout(() => {
-      setUser(null)
-      navigate('/pt/login')
-    },2000)
-    
+    postData('auth/close')
   }
   
-  return {handleOnClickCloseSession}
+  return {handleOnClickCloseSession,loading}
 }
 
 export default useCloseSession;
